Set up a config folder and move paths into it
Also move in the sass-lint configuration.
