+++ config/paths.js
@@ -1,0 +1,13 @@
+const paths = {
+  config: 'config/',
+  assets: 'app/assets/',
+  assetsScss: 'app/assets/scss/',
+  dist: 'dist/',
+  distCSS: 'dist/assets/css/',
+  distScss: 'dist/assets/scss/',
+  templates: 'app/templates/',
+  npm: 'node_modules/',
+  specs: 'test/specs/'
+}
+
+module.exports = paths
